feat: implement (incorrectly) texturing Bundle SDL3_image
Tilemaps now present the interface SetTile and GetTile and do not
expose internal map and tile details. Tiles are allocated once with
their textures and colors and are used repeatedly throughout the
whole map. Empty tiles are represented by a NULL pointer.

FIXME: Tiles are drawn textured, but the texture coordinates are not correctly
calculated. This should be fixed before merging.